fix: various changes from code review
