+++ .github/workflows/update-csv.yml
@@ -1,4 +1,4 @@
-name: Update BSB Database
+name: Update BSB Directory and Prefix Files
 
 on:
   schedule:
@@ -9,35 +9,30 @@ on:
 
 jobs:
   build:
-    name: Update BSB Database & Version
+    name: Update BSB CSVs
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
       - name: Set up Python
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: '3.x'
       - name: Install dependencies
         run: python -m pip install --upgrade pip
-      - name: Update BSB Database & Version
+      - name: Update BSB CSVs
         id: update-csvs
-        working-directory: ./src/bsb/
-        run: python ./updater/updater.py
+        run: python .github/workflows/scripts/csv_update.py
       - id: get-date
         run: echo "date=$(date +'%B-%Y')" >> $GITHUB_OUTPUT
       - uses: peter-evans/create-pull-request@v5
         id: raise-pr
         with:
-          branch: "bot/automate-sftp/${{ steps.get-date.outputs.date }}"
+          branch: "bot/automate-ftp/${{ steps.get-date.outputs.date }}"
           title: "chore: update bsb directory with ${{ steps.get-date.outputs.date }} data"
           commit-message: "chore: update bsb directory with ${{ steps.get-date.outputs.date }} data"
-          body: "Updated csv files using files: ${{ steps.update-csvs.outputs.bsb_prefix_file }}
-                  
-                  and ${{ steps.update-csvs.outputs.bsb_directory_file }}.
-                  
-                  
-                  Auto-generated by create-pull-request:
-                  [https://github.com/peter-evans/create-pull-request]"
+          body: "Updated csv files using files: ${{ steps.update-csvs.outputs.bsb_prefix_file }}\n
+                  and ${{ steps.update-csvs.outputs.bsb_directory_file }}.\n\n
+                  Auto-generated by create-pull-request:\t[https://github.com/peter-evans/create-pull-request]"
           committer: GitHub Actions <actions@github.com>
           signoff: false
           delete-branch: true
