--- scritps/auspaynet_ftp_service.py
@@ -1,26 +1,0 @@
-# from ftplib import FTP
-import pysftp
-BASE_DIR = "/~auspaynetftp/bsb"
-
-
-def fetch_latest_file(prefix: str, file_type: str) -> (bytearray, str):
-    with FTP("bsb.hostedftp.com") as ftp:
-        ftp.login()
-
-        files = ftp.mlsd(BASE_DIR)
-
-        ordered_files = sorted(files, reverse=True, key=lambda x: x[1]['modify'])
-
-        matching_filename = next(
-            (filename for filename, _ in ordered_files if
-             filename.startswith(prefix) and filename.endswith(file_type)),
-            None,
-        )
-
-        if matching_filename is None:
-            raise "Unable to find matching file"
-
-        bytes_file = bytearray()
-        ftp.retrbinary(f"RETR {BASE_DIR}/{matching_filename}", lambda data: bytes_file.extend(data))
-
-        return bytes_file, matching_filename
